refactor(accelerator): split main.rs into tray.rs and windows.rs modules
Extract tray and window management into focused binary-crate modules:

- tray.rs (176 lines): icon statics, menu building, tray construction,
  animation loop. ICON_IDLE stays private — build_tray_icon encapsulates it.
- windows.rs (104 lines): settings window, auth popup with 60s timeout,
  update prompt. focus_window stays private.
- main.rs (372 lines, down from 610): setup wiring, logging, HTTPS startup,
  update polling, exit handling. Clearly structured with section comments.

The setup closure is now readable wiring code that calls into focused
modules rather than containing all implementations inline.

No behavior change — pure code organization refactor.
